Adding variable to enforce yanda-daliuge rebuilds
Normally the GitLab CI environment will cache the layers used to build
docker images, and since we when triggering a pipeline run to produce
new images, these caches are used instead of a new image being build,
because we are not changing the Dockerfile. Adding this variable will
allow us in the future (and certainly on this commit already) to trigger
pipelines that will internally trigger an image rebuild.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>